+++ experiments/003_reproduce_lmm_cmaes/main.py
@@ -15,12 +15,12 @@ from tqdm import tqdm
 
 from optilab.data_classes import ExperimentMetadata, ExperimentResults
 from optilab.metamodels import ApproximateRankingMetamodel
-from optilab.objective_functions import ObjectiveFunction
-from optilab.objective_functions.benchmarks.cec2017_objective_function import (
+from optilab.functions import ObjectiveFunction
+from optilab.functions.benchmarks.cec2017_objective_function import (
     CEC2017ObjectiveFunction,
 )
-from optilab.objective_functions.surrogate import LocallyWeightedPolynomialRegression
-from optilab.objective_functions.unimodal import BentCigarFunction, SphereFunction
+from optilab.functions.surrogate import LocallyWeightedPolynomialRegression
+from optilab.functions.unimodal import BentCigarFunction, SphereFunction
 
 # from optilab.objective_functions.unimodal.sphere_function import SphereFunction
 from optilab.plotting import plot_ecdf_curves
