+++ src/optilab/functions/unimodal/bent_cigar_function.py
