+++ src/optilab/functions/unimodal/cumulative_squared_sums.py
